[refactor] Address 리뷰 반영 수정

+++ member/src/main/java/com/example/ordering_lecture/address/service/AddressService.java
@@ -5,9 +5,9 @@ import com.example.ordering_lecture.address.dto.AddressRequestDto;
 import com.example.ordering_lecture.address.dto.AddressResponseDto;
 import com.example.ordering_lecture.address.dto.AddressUpdateDto;
 import com.example.ordering_lecture.address.repository.AddressRepository;
+import com.example.ordering_lecture.common.ErrorCode;
+import com.example.ordering_lecture.common.OrTopiaException;
 import com.example.ordering_lecture.member.domain.Member;
-import com.example.ordering_lecture.member.dto.Buyer.MemberResponseDto;
-import com.example.ordering_lecture.member.dto.Buyer.MemberUpdateDto;
 import com.example.ordering_lecture.member.repository.MemberRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -28,13 +28,14 @@ public class AddressService {
     }
     public Address createAddress(String email, AddressRequestDto addressRequestDto) {
         Member member = memberRepository.findByEmail(email)
-                .orElseThrow(() -> new IllegalArgumentException("해당 이메일을 가진 회원이 존재하지 않습니다."));
+                .orElseThrow(() -> new OrTopiaException(ErrorCode.NOT_FOUND_MEMBER));
         Address address = addressRequestDto.toEntity(member);
         return addressRepository.save(address);
     }
 
     public AddressResponseDto findById(Long id) {
-        Address address = addressRepository.findById(id).orElseThrow();
+        Address address = addressRepository.findById(id)
+                .orElseThrow(() -> new OrTopiaException(ErrorCode.NOT_FOUND_ADDRESS));
         return AddressResponseDto.toDto(address);
     }
     public List<AddressResponseDto> showAllAddress(){
@@ -44,19 +45,27 @@ public class AddressService {
     }
     public List<AddressResponseDto> findAllAddressesByEmail(String email) {
         Member member = memberRepository.findByEmail(email)
-                .orElseThrow(() -> new IllegalArgumentException("해당 이메일을 가진 회원이 존재하지 않습니다."));
+                .orElseThrow(() -> new OrTopiaException(ErrorCode.NOT_FOUND_MEMBER));
         List<Address> addresses = addressRepository.findAllByMember(member);
+        if (addresses.isEmpty()) {
+            throw new OrTopiaException(ErrorCode.NOT_FOUND_ADDRESS_BY_EMAIL);
+        }
         return addresses.stream()
                 .map(AddressResponseDto::toDto)
                 .collect(Collectors.toList());
     }
     @Transactional
     public AddressResponseDto updateAddress(Long id, AddressUpdateDto addressUpdateDto) {
-        Address address = addressRepository.findById(id).orElseThrow();
+        Address address = addressRepository.findById(id)
+                .orElseThrow(() -> new OrTopiaException(ErrorCode.NOT_FOUND_ADDRESS));
         address.updateAddress(addressUpdateDto);
         return AddressResponseDto.toDto(address);
     }
+
     public void deleteAddress(Long id) {
+        if (!addressRepository.existsById(id)) {
+            throw new OrTopiaException(ErrorCode.NOT_FOUND_ADDRESS);
+        }
         addressRepository.deleteById(id);
     }
 }
